+++ __mocks__/@redocly/openapi-core.ts
@@ -1,14 +1,16 @@
 export const __redoclyClient = {
-  isAuthorizedWithRedocly: jest.fn(() => true),
+  isAuthorizedWithRedocly: jest.fn().mockResolvedValue(true),
+  isAuthorizedWithRedoclyByRegion: jest.fn().mockResolvedValue(true),
   login: jest.fn(),
-  getOrganizationId: jest.fn(() => ({ organizationById: 'organizationById' })),
-  getDefinitionVersion: jest.fn(),
-  getSignedUrl: jest.fn(() => ({
-    signFileUploadCLI: { signedFileUrl: null, uploadedFilePath: null },
-  })),
-  getDefinitionByName: jest.fn(() => ({ definition: 'definition' })),
-  createDefinitionVersion: jest.fn(),
-  updateDefinitionVersion: jest.fn(),
+  registryApi: {
+    setAccessTokens: jest.fn(),
+    authStatus: jest.fn(),
+    prepareFileUpload: jest.fn().mockResolvedValue({
+      signedUploadUrl: 'signedUploadUrl',
+      filePath: 'filePath',
+    }),
+    pushApi: jest.fn(),
+  },
 };
 
 export const RedoclyClient = jest.fn(() => __redoclyClient);
@@ -20,3 +22,4 @@ export const lint = jest.fn();
 export const bundle = jest.fn(() => ({ bundle: { parsed: null }, problems: null }));
 export const getTotals = jest.fn(() => ({ errors: 0 }));
 export const formatProblems = jest.fn();
+export const slash = jest.fn();