+++ __tests__/bundle/max-problems-argument/snapshot.js
@@ -236,6 +236,8 @@ components:
       items:
         $ref: '#/components/schemas/Pet'
 
+No configurations were defined in extends -- using built in recommended configuration by default.
+
 foo.yaml:
   29:11  error    spec  Property \`header\` is not expected here.
 
@@ -244,6 +246,8 @@ run \`openapi lint --generate-ignore-file\` to add all problems to the ignore fi
 
 bundling ./foo.yaml...
 📦 Created a bundle for ./foo.yaml at stdout <test>ms.
+No configurations were defined in extends -- using built in recommended configuration by default.
+
 bar.yaml:
   14:7  error    spec  Property \`summay\` is not expected here.