+++ packages/core/src/rules/common/__tests__/paths-kebab-case.test.ts
@@ -83,4 +83,27 @@ describe('Oas3 paths-kebab-case', () => {
       ]
     `);
   });
+
+  it('should allow trailing slash in path with "paths-kebab-case" rule', async () => {
+      const document = parseYamlToDocument(
+        outdent`
+            openapi: 3.0.0
+            paths:
+              /some/:
+                get:
+                  summary: List all pets
+          `,
+        'foobar.yaml',
+      );
+
+      const results = await lintDocument({
+        externalRefResolver: new BaseResolver(),
+        document,
+        config: makeConfig({
+          'paths-kebab-case': 'error',
+          'no-path-trailing-slash': 'off',
+        }),
+      });
+      expect(replaceSourceWithRef(results)).toMatchInlineSnapshot(`Array []`);
+    });
 });